requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
google-cloud-datastore==1.7.3
google-auth==1.6.2
requests==2.21.0
